Introduz conceito de RideStorage (storage carona)
É possível iniciar um RideStorage  já com uma IQueryable<TDataModel> definida, assim
podemos utilizar os métodos auxiliares de um IStorageReader padrão
